pyproject.toml
readme = "README.md"

[tool.poetry]
name = "riscv_watermark"
version = "0.0.1"

packages = [
    { include = "riscv_watermark"},
]

[tool.poetry.scripts]
riscv-watermark = "riscv_watermark.main:main"

[tool.poetry.dependencies]
python = "^3.13"
pyelftools = "^0.31"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"